pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai_manga_workflow"
version = "0.1.0"
description = "A fully automated AI Manga generation workflow system"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "python-dotenv",
    "requests",
    "veadk",
    "google-adk",
    "tos",
    "mcp",
    "imageio-ffmpeg",
]

[project.urls]
Homepage = "https://github.com/yourusername/ai_manga_workflow"

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]
exclude = ["tests*", "img*"]
